routing compression

+++ test.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ "x$1" = "xhelp" ]; then
+  MAX_LIMIT=`ls -d 0* | tail -n 1 |awk -F- '{print $1}' | sed -e "s#^0*##g"`
+  echo "[DELAYEVALUATION=1] [LIMIT=[1-$MAX_LIMIT]] $0"
+  exit 0
+fi
+
 rm -f testbed.pdf
 
 if [ "x$VALGRIND" = "x" ]; then
@@ -32,12 +38,11 @@ if [ "x$START" = "x" ]; then
   START=1
 fi
 
-
 NO_SIM=`echo $SIMULATOR | wc -w`
 
 for next_sim in $SIMULATOR; do
 
-  SIMULATOR=$next_sim LATEX=$ENABLE_LATEX NOSUPP=$NOSUPP VALGRINDOPTION="$VALGRINDOPTION" VALGRIND=$VALGRIND_PARAMS VALGRINDXML=$VALGRINDXML_PARAMS MODE=SIMULATION START=$START LIMIT=$LIMIT ./tools/testing/gen_test_pdf.sh
+  SIMULATOR=$next_sim LATEX=$ENABLE_LATEX NOSUPP=$NOSUPP VALGRINDOPTION="$VALGRINDOPTION" VALGRIND=$VALGRIND_PARAMS VALGRINDXML=$VALGRINDXML_PARAMS MODE=SIMULATION START=$START LIMIT=$LIMIT DELAYEVALUATION=1 ./tools/testing/gen_test_pdf.sh
 
   if [ -f testbed.pdf ] && [ $NO_SIM -gt 1 ]; then
     mv testbed.pdf testbed_$next_sim.pdf
